Set argument types to fix issue #25
Explicitly set the argument and return types of the clipboard functions used in the module. This fixed the error described in issue #25 on my computer. The problem seems to be due to larger 64-bit pointers and handles used while running in 64 bits on 64-bit Windows, as opposed to the smaller size ints that ctypes uses by default.