Introduced lazy loading, so that modules aren't loaded immediately after importing pyperclip so the user can call set_clipboard() first.
Changed behavior so that copy() and paste() are set to lazy_load_stub_copy() and lazy_load_stub_paste() until the first time copy(), paste(), determine_clipboard(), or set_clipboard() are called. Originally copy() and paste() were set to None, which is how the user could determine if a clipboard mechanism succeeded or not. This is a backwards-incompatible change.

The is_available() function can determine if a clipboard mechanism was loaded or not.

Renamed the init_osx_cmd_clipboard() function to init_osx_pbcopy_clipboard().

Added a warning for cygwin, since there's no implementation for it yet.

Added a ValueError if set_clipboard() wasn't passed a valid clipboard mechanism.

Changed __all__ to include determine_clipboard() and set_clipboard()

Added more comments and documentation in general.

Renamed text_type constant to STR_OR_UNICODE.