Use distutils.spawn.find_executable instead
stdout and stderr may not always be available. The most common case of this is on Windows when using pythonw.exe instead of python.exe. This method is therefore a safer method, especially on Windows, to determine if an executable exists. See https://docs.python.org/3/library/sys.html#sys.__stderr__ for more info