better wayland session detection
XDG_SESSION_TYPE may not always be set to "wayland" or "x11" depending
on what compositor/setup the user is using. For instance, weston (the
reference compositor) does not set this particular environment variable.
Instead, check for the existence of WAYLAND_DISPLAY. This is set by all
wayland compositors and tells clients which wayland compositor to
connect to. If we have this variable set, we can be sure wayland is
being used.