Bumping version to 1.9.0.
Automatically cast the copy() argument to a string for all data types.

_py3_executable_exists and _py2_executable_exists had swapped names; fixed.

Pyperclip now "stringifies" all data types by passing it to str() (or globals()['__builtins__'].unicode on Python 2), so passing [1, 2, 3] would put '[1, 2, 3]' on the clipboard.

shutil.which() replaces the custom code (except in 2.7 and below which doesn't have shutil.which()).

Remove waitForPaste() and waitForNewPaste() functions, these aren't something the core library should have.

Reordered so that xclip is chosen before xsel since xclip is more popular.

Removed "where" command for Windows because we don't need it on Windows.